Moved CsvParser into its own file
